requirements.txt
# alabaster==0.7.12
amqp==2.4.2
# apistar==0.7.1
argon2-cffi==18.3.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.3
Babel==2.6.0
backcall==0.1.0
backports.csv==1.0.7
beautifulsoup4==4.8.2
billiard==3.5.0.5
bleach==3.1.0
# bootstrap-admin==0.4.3
celery==4.2.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
decorator==4.4.0
defusedxml==0.5.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2
django-admin-honeypot==1.1.0
django-admin-interface==0.12.0
django-admin-timeline==1.7.2
django-admin-tools==0.8.1
django-adminlte-ui==1.4.0
django-adminlte3==0.1.0
django-appconf==1.0.3
django-autoslug-iplweb==1.9.5
django-avatar==4.1.0
django-bootstrap4==1.1.1
django-colorfield==0.1.15
django-controlcenter==0.2.9
django-cors-headers==3.2.0
django-coverage-plugin==1.6.0
django-crispy-forms==1.7.2
django-dash==0.5.6
django-dashing==0.3.4
django-debug-toolbar==2.1
django-dynamic-preferences==1.8.1
django-environ==0.4.5
django-extensions==2.2.5
django-filter==2.2.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-fluent-dashboard==1.0.1
django-formtools==2.1
django-grappelli==2.13.3
django-image-cropping==1.2.0
django-imagekit==4.0.2
-e git+https://github.com/django-import-export/django-import-export.git@aa9284fac637b461bb9657270ef2df6f17ca1800#egg=django_import_export
django-js-asset==1.2.2
django-json-widget==0.2.0
django-jsoneditor==0.1.4
django-model-utils==3.1.2
django-mptt==0.10.0
django-mptt-admin==0.6.0
-e git://github.com/theatlantic/django-nested-admin.git@23d1faa51f37d8e055dcee4c6d66cdceb0a323c6#egg=django_nested_admin
django-nested-inline==0.4.0
django-nine==0.2.3
django-nucleus==0.1.2
django-nyt==1.1.4
django-pagedown==1.0.6
django-pkgconf==0.3.0
django-ranged-response==0.2.0
django-redis==4.9.0
django-reverse-admin==2.8.1
django-reversion==3.0.4
django-select2==6.3.1
django-simple-captcha==0.5.12
django-simple-history==2.8.0
django-simpleui==3.9
django-tinymce==2.8.0
django-treebeard==4.3
django-vali==0.2.0
djangorestframework==3.10.3
djangorestframework-camel-case==1.1.2
djangorestframework-filters==1.0.0.dev0
djangorestframework-recursive==0.1.2
docopt==0.6.2
docutils==0.14
drf-flex-fields==0.6.1
drf-nested-routers==0.91
drf-yasg==1.17.0
dynamic-rest==1.8.1
easy-thumbnails==2.7
entrypoints==0.3
et-xmlfile==1.0.1
factory-boy==2.11.1
Faker==1.0.4
feedparser==5.2.1
ffmpeg-python==0.1.17
flake8==3.5.0
flower==0.9.2
future==0.15.2
httplib2==0.9.2
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.2.0
imagesize==1.1.0
inflection==0.3.1
ipdb==0.11
ipykernel==5.1.2
ipython==7.3.0
ipython-genutils==0.2.0
isort==4.3.15
itypes==1.1.0
jdcal==1.4
jedi==0.13.3
jet-django==0.7.4
Jinja2==2.10
json5==0.8.5
jsonschema==3.0.2
jupyter-client==5.3.1
jupyter-core==4.5.0
jupyterlab==1.1.0
jupyterlab-server==1.0.6
kombu==4.4.0
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==6.0.0
moviepy==1.0.0
mypy==0.641
mypy-extensions==0.4.1
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.16.2
odfpy==1.4.0
openpyxl==2.6.1
ordereddict==1.1
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
pendulum==2.0.4
persisting-theory==0.2.1
pickleshare==0.7.5
pif==0.8.2
pilkit==2.0
Pillow==5.3.0
pipreqs==0.4.10
pluggy==0.9.0
proglog==0.1.9
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2==2.7.5
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==1.6.0
Pygments==2.3.1
pylint==2.3.1
pylint-django==2.0.13
pylint-plugin-utils==0.6
pyparsing==2.3.1
PyPDF2==1.26.0
pyrsistent==0.15.4
pytest==3.9.1
pytest-django==3.4.3
pytest-sugar==0.9.1
python-dateutil==2.8.0
python-monkey-business==1.0.0
python-slugify==1.2.6
pytz==2018.5
pytzdata==2018.9
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.1.0
redis==3.2.1
requests==2.21.0
ruamel.yaml==0.15.89
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==1.2.1
soupsieve==1.9.5
Sphinx==1.8.1
sphinxcontrib-websupport==1.1.0
sqlparse==0.3.0
tablib==0.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
tornado==6.0.3
tqdm==4.31.1
traitlets==4.3.2
transliterate==1.10.2
treebeard==0.1.1
typed-ast==1.3.1
typesystem==0.2.0
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
vishap==0.1.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==4.1
wrapt==1.11.1
xlrd==1.2.0
xlwt==1.3.0
yarg==0.1.9
